Restructure the directory tree
Restructure the tree by doing:

- Move `protocols` to be a submodule of `channel`
- Move integration tests to `tests` in the project root directory
- Put internal `tests` module under `channel` module

Please note: The code in `src/channel/tests/harness` is a subset cut'n' pasta
from the integration tests. While this means we have 300 lines of duplicate code
this is much cleaner than any of the other workarounds required to get `rustc`
to play nicely while still enforcing integration tests can only test the public
API.

With this applied we get the directory tree:

    .
    ├── Cargo.toml
    ├── README.md
    ├── src
    │   ├── channel
    │   │   ├── protocols
    │   │   │   ├── close.rs
    │   │   │   ├── create.rs
    │   │   │   ├── punish.rs
    │   │   │   ├── splice.rs
    │   │   │   └── update.rs
    │   │   ├── protocols.rs
    │   │   ├── tests
    │   │   │   └── harness
    │   │   │       ├── mod.rs
    │   │   │       └── wallet.rs
    │   │   └── tests.rs
    │   ├── channel.rs
    │   ├── keys.rs
    │   ├── lib.rs
    │   ├── serde
    │   │   └── partially_signed_transaction.rs
    │   ├── serde.rs
    │   ├── signature.rs
    │   ├── transaction
    │   │   └── ptlc.rs
    │   └── transaction.rs
    └── tests
        ├── e2e.rs
        └── harness
            ├── mod.rs
            └── wallet.rs
