+++ thor/src/channel/tests.rs
@@ -1,8 +1,71 @@
-use crate::{
-    test_harness::{create_channels, init_bitcoind, init_cli, swap_beta_ptlc_bob, FUND},
-    MedianTime, PtlcSecret, TX_FEE,
-};
+pub mod harness;
+
+use crate::{MedianTime, PtlcSecret, TX_FEE};
+use harness::{create_channels, init_bitcoind, init_cli, swap_beta_ptlc_bob, FUND, update_balances};
+
 use bitcoin::Amount;
+
+#[tokio::test]
+async fn e2e_punish_publication_of_revoked_commit_transaction() {
+    let cli = init_cli();
+    let bitcoind = init_bitcoind(&cli).await;
+    let (
+        mut a_channel,
+        mut b_channel,
+        mut a_transport,
+        mut b_transport,
+        a_wallet,
+        b_wallet,
+        time_lock,
+        _,
+    ) = create_channels(&bitcoind).await;
+
+    let a_balance_after_open = a_wallet.balance().await.unwrap();
+    let b_balance_after_open = b_wallet.balance().await.unwrap();
+
+    // Parties agree on a new channel balance: Alice pays 0.5 a Bitcoin to Bob
+    let payment = Amount::from_btc(0.5).unwrap();
+    let a_balance = FUND - payment;
+    let b_balance = FUND + payment;
+
+    update_balances(
+        &mut a_channel,
+        &mut b_channel,
+        &mut a_transport,
+        &mut b_transport,
+        a_balance,
+        b_balance,
+        time_lock,
+    )
+    .await;
+
+    // Alice attempts to cheat by publishing a revoked commit transaction.
+    let signed_revoked_tx_c = a_channel.latest_revoked_signed_tx_c().unwrap().unwrap();
+    a_wallet
+        .0
+        .send_raw_transaction(signed_revoked_tx_c.clone())
+        .await
+        .unwrap();
+
+    // Bob sees the transaction and punishes Alice.
+    b_channel
+        .punish(&b_wallet, signed_revoked_tx_c)
+        .await
+        .unwrap();
+
+    let a_balance_after_punish = a_wallet.balance().await.unwrap();
+    let b_balance_after_punish = b_wallet.balance().await.unwrap();
+
+    assert_eq!(
+        a_balance_after_punish, a_balance_after_open,
+        "Alice should get no money back after being punished"
+    );
+    assert_eq!(
+        b_balance_after_punish,
+        b_balance_after_open + FUND * 2 - Amount::from_sat(TX_FEE) * 2,
+        "Bob should get all the money back after punishing Alice"
+    );
+}
 
 #[tokio::test]
 async fn bob_can_refund_ptlc_if_alice_holds_onto_secret_after_first_update() {
