+++ thor/src/channel/tests/harness/wallet.rs
@@ -1,13 +1,15 @@
 use crate::{
-    BroadcastSignedTransaction, BuildFundingPsbt, GetRawTransaction, MedianTime, NewAddress,
-    SignFundingPsbt,
+    channel::{BuildFundingPsbt, SignFundingPsbt},
+    BroadcastSignedTransaction, GetRawTransaction, MedianTime, NewAddress,
 };
+
 use anyhow::Result;
 use async_trait::async_trait;
-use bitcoin::{consensus, util::psbt::PartiallySignedTransaction, Address, Amount};
+use bitcoin::{util::psbt::PartiallySignedTransaction, Address, Amount};
 use bitcoin_harness::{bitcoind_rpc::PsbtBase64, Bitcoind};
 use reqwest::Url;
 
+#[derive(Debug)]
 pub struct Wallet(pub bitcoin_harness::Wallet);
 
 impl Wallet {
@@ -16,16 +18,16 @@ impl Wallet {
 
         Ok(Self(wallet))
     }
-
+    
     pub async fn balance(&self) -> Result<Amount> {
-        let b = self.0.balance().await?;
-        Ok(b)
+        let balance = self.0.balance().await?;
+        Ok(balance)
     }
 }
 
-/// Create two bitcoind wallets on the `bitcoind` node and fund them with the
-/// amount that they will contribute to the channel, plus a little extra to
-/// cover transaction fees.
+/// Create two bitcoind wallets on the node passed as an argument and fund them
+/// with the amount that they will contribute to the channel, plus a buffer to
+/// account for transaction fees.
 pub async fn make_wallets(
     bitcoind: &Bitcoind<'_>,
     fund_amount: Amount,
@@ -36,12 +38,18 @@ pub async fn make_wallets(
     Ok((alice, bob))
 }
 
-async fn make_wallet(name: &str, bitcoind: &Bitcoind<'_>, fund_amount: Amount) -> Result<Wallet> {
+async fn make_wallet(
+    name: &str,
+    bitcoind: &Bitcoind<'_>,
+    fund_amount: Amount,
+) -> Result<Wallet> {
     let wallet = Wallet::new(name, bitcoind.node_url.clone()).await?;
-    let address = wallet.0.new_address().await.unwrap();
-    let extra = Amount::from_btc(1.0)?;
+    let buffer = Amount::from_btc(1.0).unwrap();
+    let amount = fund_amount + buffer;
 
-    bitcoind.mint(address, fund_amount + extra).await.unwrap();
+    let address = wallet.0.new_address().await.unwrap();
+
+    bitcoind.mint(address, amount).await.unwrap();
 
     Ok(wallet)
 }
@@ -55,7 +63,8 @@ impl BuildFundingPsbt for Wallet {
     ) -> Result<PartiallySignedTransaction> {
         let psbt = self.0.fund_psbt(output_address, output_amount).await?;
         let as_hex = base64::decode(psbt)?;
-        let psbt = consensus::deserialize(&as_hex)?;
+
+        let psbt = bitcoin::consensus::deserialize(&as_hex)?;
 
         Ok(psbt)
     }
@@ -67,14 +76,14 @@ impl SignFundingPsbt for Wallet {
         &self,
         psbt: PartiallySignedTransaction,
     ) -> Result<PartiallySignedTransaction> {
-        let psbt = consensus::serialize(&psbt);
+        let psbt = bitcoin::consensus::serialize(&psbt);
         let as_base64 = base64::encode(psbt);
 
         let psbt = self.0.wallet_process_psbt(PsbtBase64(as_base64)).await?;
         let PsbtBase64(signed_psbt) = PsbtBase64::from(psbt);
 
         let as_hex = base64::decode(signed_psbt)?;
-        let psbt = consensus::deserialize(&as_hex)?;
+        let psbt = bitcoin::consensus::deserialize(&as_hex)?;
 
         Ok(psbt)
     }
@@ -88,7 +97,7 @@ impl BroadcastSignedTransaction for Wallet {
         // TODO: Instead of guessing how long it will take for the transaction to be
         // mined we should ask bitcoind for the number of confirmations on `txid`
 
-        // Give time for transaction to be mined.
+        // give time for transaction to be mined
         tokio::time::delay_for(std::time::Duration::from_millis(1100)).await;
 
         Ok(())
@@ -102,19 +111,16 @@ impl NewAddress for Wallet {
     }
 }
 
-#[async_trait::async_trait]
+#[async_trait]
 impl MedianTime for Wallet {
-    async fn median_time(&self) -> anyhow::Result<u32> {
+    async fn median_time(&self) -> Result<u32> {
         self.0.median_time().await.map_err(Into::into)
     }
 }
 
-#[async_trait::async_trait]
+#[async_trait]
 impl GetRawTransaction for Wallet {
-    async fn get_raw_transaction(
-        &self,
-        txid: bitcoin::Txid,
-    ) -> anyhow::Result<bitcoin::Transaction> {
+    async fn get_raw_transaction(&self, txid: bitcoin::Txid) -> Result<bitcoin::Transaction> {
         self.0.get_raw_transaction(txid).await.map_err(Into::into)
     }
 }
